requirements.txt
sentence-transformers>=3.0.0
faiss-cpu
datasets
fastapi
uvicorn[standard]
scikit-learn
google-cloud-firestore
pydantic
numpy
tqdm
requests
pandas
joblib
google-generativeai
pinecone>=3.0.0
sentence-transformers>=2.2.2
numpy>=1.23
sentence-transformers>=2.2.2
numpy>=1.23
